+++ src/mysqlsh/history.cc
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, 2018, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2017, 2019, Oracle and/or its affiliates. All rights reserved.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License, version 2.0,
@@ -134,7 +134,7 @@ void History::clear_temporary() {
 }
 
 void History::set_limit(uint32_t limit) {
-  if (limit > std::numeric_limits<int>::max() - 1)
+  if (limit > static_cast<uint32_t>(std::numeric_limits<int>::max() - 1))
     throw std::invalid_argument("invalid history limit");
   // Note: the linenoise history includes the "current" line being edited, thus
   // it can't be set to 0. OTOH users are not expecting that line to be counted
